!B (Audio) False error warnings, because of empty connections of "do_nothing" trigger. Added class CDoNothingTrigger and execute triggers on CATLTriggerImpl instead of CATLAudioObject to prevent this. !B (Audio) Added missing warnings for duplicated switches and parameters. !B (Audio) Added missing warnings about unknown XML tags to middleware implementations. !XB (Audio) Default controls library was not marked as modified after migration to new ACE structure. Review: ID 69787 by Author: Marius Popczyk, Reviewer: Thomas Wollenzin, Observer: Luis Lairet  (Approved by thomasw)
Copied from Perforce
 Change: 1634885

+++ Code/Sandbox/Plugins/EditorAudioControlsEditor/SystemAssets.h
@@ -125,7 +125,7 @@ public:
 	void                    RemoveConnection(ConnectionPtr const pConnection);
 	void                    RemoveConnection(CImplItem* const pImplControl);
 	void                    ClearConnections();
-	ConnectionPtr           GetConnectionAt(int const index) const;
+	ConnectionPtr           GetConnectionAt(size_t const index) const;
 	ConnectionPtr           GetConnection(CID const id) const;
 	ConnectionPtr           GetConnection(CImplItem const* const pImplControl) const;
 	void                    ReloadConnections();
